Added slider functionality
Added simple jQuery slider functionality on the front gallery after
experiencing lots of trouble trying to get the Nivo slider integrated
into the grid gallery page. This does not include any left or right
buttons for switching images at the moment. Not sure if they will be
necessary.

Added jQuery code to remove the slider when the ajax request is made to
show additional images when clicking on one of the grid gallery images.
The slider will reappear when the use returns to the main page.

Added reload on orientation change for mobile devices that are rotated.
This is necessary because the size of the slider image varies according
to the size of the users screen. There is a height that is added to the
div containing the slider so that the grid gallery does not move when
the slider is finally loaded into place. If the screen is rotated, the
size of the div containing the slider needs to be re-evaluated and set
appropriately for the new landscape or portrait orientation.